Version 0.7.0
Removed some unnecessary variables
Replaced int2str with std::to_string
Fixed problem with powerups after client disconnected
Fixed problem when accesing a removed socket after client disconnected
Fixed ready not showing up at clients
Implemented darkness
Added random angle to glitch
General cleanup

+++ classes/renderer.h
@@ -11,14 +11,14 @@
 #include "Game.h"
 #include "Player.h"
 #include "Client.h"
-#include "Object.h"
+#include "ObjectBase.h"
 //
 class Renderer{
     public:
         // Vars
         sf::RenderWindow window;
         std::thread renderThread;
-        float fps;
+        int fps;
         // Functions
         void start(Main &main, const Config &config,Game &game,std::vector<std::unique_ptr<ObjectBase>> &objects, const std::vector<Player> &player);
         void thread(Main &main, const Config &config,Game &game,std::vector<std::unique_ptr<ObjectBase>> &objects, const std::vector<Player> &player);
